+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to
 ### Added
 
 - cosmwasm-std: Add `Uint{64,128,256,512}::one`.
+- cosmwasm-std: Add `Uint{64,128,256,512}::abs_diff` and
+  `Decimal{,256}::abs_diff` ([#1334]).
+
+[#1334]: https://github.com/CosmWasm/cosmwasm/pull/1334
 
 ## [1.0.0] - 2022-05-14